+++ packages/core-data/src/entity-types/index.ts
@@ -5,6 +5,12 @@ import type { Context, Updatable } from './helpers';
 import type { Attachment } from './attachment';
 import type { Base, TemplatePartArea, TemplateType } from './base';
 import type { Comment } from './comment';
+import type {
+	FontCollection,
+	CollectionFontFamily,
+	CollectionFontFace,
+} from './font-collection';
+import type { FontFamily, FontFace, WpFontFamily } from './font-family';
 import type { GlobalStylesRevision } from './global-styles-revision';
 import type { MenuLocation } from './menu-location';
 import type { NavMenu } from './nav-menu';
@@ -31,8 +37,13 @@ export type { BaseEntityRecords } from './base-entity-records';
 export type {
 	Attachment,
 	Base as UnstableBase,
+	CollectionFontFace,
+	CollectionFontFamily,
 	Comment,
 	Context,
+	FontCollection,
+	FontFace,
+	FontFamily,
 	GlobalStylesRevision,
 	MenuLocation,
 	NavMenu,
@@ -54,6 +65,7 @@ export type {
 	User,
 	Widget,
 	WidgetType,
+	WpFontFamily,
 	WpTemplate,
 	WpTemplatePart,
 };
@@ -95,6 +107,7 @@ export interface PerPackageEntityRecords< C extends Context > {
 		| Base< C >
 		| Attachment< C >
 		| Comment< C >
+		| FontCollection< C >
 		| GlobalStylesRevision< C >
 		| MenuLocation< C >
 		| NavMenu< C >
@@ -113,6 +126,7 @@ export interface PerPackageEntityRecords< C extends Context > {
 		| Type< C >
 		| Widget< C >
 		| WidgetType< C >
+		| WpFontFamily< C >
 		| WpTemplate< C >
 		| WpTemplatePart< C >;
 }