+++ packages/dataviews/src/field-types/password.tsx
@@ -1,24 +1,14 @@
 /**
  * Internal dependencies
  */
-import type {
-	DataViewRenderFieldProps,
-	Field,
-	NormalizedField,
-	Rules,
-	SortDirection,
-} from '../types';
+import type { DataViewRenderFieldProps } from '../types';
+import type { FieldType } from '../types/private';
 import RenderFromElements from './utils/render-from-elements';
-import { getControl } from '../dataform-controls';
-import hasElements from './utils/has-elements';
-import getValueFromId from './utils/get-value-from-id';
-import setValueFromId from './utils/set-value-from-id';
-
-/* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-function sort( _valueA: any, _valueB: any, _direction: SortDirection ) {
-	// Passwords should not be sortable for security reasons
-	return 0;
-}
+import isValidRequired from './utils/is-valid-required';
+import isValidMinLength from './utils/is-valid-min-length';
+import isValidMaxLength from './utils/is-valid-max-length';
+import isValidPattern from './utils/is-valid-pattern';
+import isValidElements from './utils/is-valid-elements';
 
 function render( { item, field }: DataViewRenderFieldProps< any > ) {
 	return field.hasElements ? (
@@ -28,41 +18,21 @@ function render( { item, field }: DataViewRenderFieldProps< any > ) {
 	);
 }
 
-export default function normalizeField< Item >(
-	field: Field< Item >
-): NormalizedField< Item > {
-	const getValue = field.getValue || getValueFromId( field.id );
-	const setValue = field.setValue || setValueFromId( field.id );
-	const isValid: Rules< Item > = {
-		elements: true,
-		custom: () => null,
-	};
-
-	return {
-		id: field.id,
-		type: 'password',
-		label: field.label || field.id,
-		header: field.header || field.label || field.id,
-		description: field.description,
-		placeholder: field.placeholder,
-		getValue,
-		setValue,
-		elements: field.elements,
-		getElements: field.getElements,
-		hasElements: hasElements( field ),
-		render: field.render ?? render,
-		Edit: getControl( field, 'password' ),
-		sort: field.sort ?? sort,
-		isValid: {
-			...isValid,
-			...field.isValid,
-		},
-		isVisible: field.isVisible,
-		enableSorting: field.enableSorting ?? false,
-		enableGlobalSearch: field.enableGlobalSearch ?? false,
-		enableHiding: field.enableHiding ?? true,
-		readOnly: field.readOnly ?? false,
-		filterBy: false,
-		format: {},
-	};
-}
+export default {
+	type: 'password',
+	render,
+	Edit: 'password',
+	sort: () => 0, // Passwords should not be sortable for security reasons
+	enableSorting: false,
+	enableGlobalSearch: false,
+	defaultOperators: [],
+	validOperators: [],
+	getFormat: () => ( {} ),
+	validate: {
+		required: isValidRequired,
+		pattern: isValidPattern,
+		minLength: isValidMinLength,
+		maxLength: isValidMaxLength,
+		elements: isValidElements,
+	},
+} satisfies FieldType< any >;