+++ routes/template-part-list/view-utils.ts
@@ -79,6 +79,20 @@ export const DEFAULT_VIEWS: {
 		},
 	},
 	{
+		slug: 'overlay',
+		label: 'Overlays',
+		view: {
+			...DEFAULT_VIEW,
+			filters: [
+				{
+					field: 'area',
+					operator: 'is',
+					value: 'overlay',
+				},
+			],
+		},
+	},
+	{
 		slug: 'uncategorized',
 		label: 'General',
 		view: {